Refactor mask
Rather than having the mask on the parent element, we put both the mask and the sidebar as children of the same div.  This is because we don’t want the sidebar to fade in as it enters, and since the mask needs to fade in, putting the sidebar as a child of the mask means the sidebar must fade in as well.  Making them siblings instead allows them to have separate animations.